reorganize a lot of sdk public interface
